Remove issues from README, Add github address
Issues were listed in the README. These were added as issues on Github,
so there is no reason to track them here any more.

Also as a part of putting the project on Github, I added the Github
project address to the gemspec file instead of pointing to my blog post.